requirements.txt
编辑器 pycharm 2023.2
python 3.11

numpy==1.25.2
Pillow==10.0.0
Pillow==10.0.1
